+++ src/onegov.gazette/onegov/gazette/utils/__init__.py
@@ -1,5 +1,7 @@
+from onegov.gazette.utils.orm import bool_is
 from onegov.gazette.utils.sogc_importer import SogcImporter
 
 __all__ = [
-    'SogcImporter'
+    'bool_is',
+    'SogcImporter',
 ]